AI able to team up with Leaders. Adding New Custom System Groups to organize Entity manager and decouple code.

+++ Assets/Scripts/IAUS/ECS Take2/Component/Attackable.cs
@@ -28,10 +28,10 @@ namespace InfluenceMap.Factions
     public struct FactionAggression
     {
         [Range(-10,10)]
-        public float Humans;
+        public int Humans;
         [Range(-10, 10)]
-        public float Daemons;
+        public int Daemons;
         [Range(-10, 10)]
-        public float Angels;
+        public int Angels;
     }
 }
